# shared/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: GNSS Flow API
  version: 0.1.0
paths:
...
  /projects/{project_id}:
    get:
      summary: Get project
      responses:
        "200":
          description: OK
        "404":
          description: Not found
    put:
      summary: Update project
      responses:
        "200":
          description: OK
        "404":
          description: Not found
    delete:
      summary: Delete project
      responses:
        "200":
          description: OK
        "404":
          description: Not found
...
  /observations/rinex:
    get:
      summary: Download recorded observations as RINEX 3.04 file
      parameters:
        - name: marker
          in: query
          schema:
            type: string
            default: GNSSFLOW
      responses:
        "200":
          description: RINEX file content
        "404":
          description: No epochs recorded
...